car-deal-bot-project/requirements.txt
requests>=2.31.0
feedparser>=6.0.11
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
twilio>=9.0.0
